refactor(syscalls): move logic into core crate
